requirements.txt
flask
<<<<<<< HEAD
tensorflow-cpu 
=======
tensorflow-cpu
>>>>>>> cec7369 (how are you)
opencv-python-headless
numpy
gunicorn
